third_party/blink/web_tests/editing/selection/iframe-select-expected.txt

CONSOLE ERROR: line 6: Uncaught IndexSizeError: Failed to execute 'setBaseAndExtent' on 'Selection': There is no child at offset 1.
